fix(M1-STATE-04/PROFILE-04): code-review 修复
/code-review 对抗审查（7 finder + 验证）后修复 3 类正确性/设计符合性问题：

- blocked 恢复强校验 target == pre_blocked_status（设计 09-appendix 行 207/218
  「恢复到进入 blocked 前的状态」）：之前接受任意 blockable 目标会丢失断点续传锚点。
  未记录 pre_blocked_status 时退化为只校验 blockable。
- 显式 --to 一律走 can_transition_to（矩阵无自环）：移除 target==from 同态短路，
  使同态/终态 --to（如对 done 模块 --to done）正确报 InvalidTransition，
  杜绝绕过终态保护、向终结模块写伪审计记录。幂等的「仅更新 substatus」走 to==None。
- parse_triple 版本段 u32 → u64：避免 CalVer/超长版本段溢出被误判「版本不足」。

新增 2 回归测试（错误恢复目标拒绝、同态 --to 拒绝）。
受控输入下低危项（版本串首个点分启发式、--version 非零退出判缺失、prerelease
后缀、序列化 null 降级）评估为可接受，未改。

质量门：fmt + clippy -D warnings + 全量测试（137 core + 23 e2e）全过。

+++ cli/crates/core/src/profile/tools.rs
@@ -173,17 +173,20 @@ fn parse_version_str(s: &str) -> Option<String> {
     None
 }
 
-/// 解析 `major.minor.patch`（缺失部分补 0）为 `(u32, u32, u32)`。
-fn parse_triple(v: &str) -> Option<(u32, u32, u32)> {
+/// 解析 `major.minor.patch`（缺失部分补 0）为 `(u64, u64, u64)`。
+///
+/// 用 u64 避免 CalVer/超长版本段（如 `20240131`）或异常输出溢出 u32 导致解析失败、
+/// 进而被 [`version_satisfies`] 保守误判为「版本不足」。
+fn parse_triple(v: &str) -> Option<(u64, u64, u64)> {
     let mut parts = v.split('.');
-    let major = parts.next()?.parse::<u32>().ok()?;
+    let major = parts.next()?.parse::<u64>().ok()?;
     let minor = parts
         .next()
-        .and_then(|p| p.parse::<u32>().ok())
+        .and_then(|p| p.parse::<u64>().ok())
         .unwrap_or(0);
     let patch = parts
         .next()
-        .and_then(|p| p.parse::<u32>().ok())
+        .and_then(|p| p.parse::<u64>().ok())
         .unwrap_or(0);
     Some((major, minor, patch))
 }
